1.5.2－3リリース
モデル関係の調整。

+++ MMMLib/net/minecraft/src/MMM_ModelMultiMMMBase.java
@@ -16,9 +16,9 @@ public abstract class MMM_ModelMultiMMMBase extends MMM_ModelMultiBase {
 	@Deprecated
 	public float onGround;
 	@Deprecated
-	public int heldItemLeft;
+	public float heldItemLeft;
 	@Deprecated
-	public int heldItemRight;
+	public float heldItemRight;
 
 
 	public MMM_ModelMultiMMMBase() {
